message bug
